% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminerWHM-package.R
\docType{package}
\name{libminerWHM-package}
\alias{libminerWHM}
\alias{libminerWHM-package}
\title{libminerWHM: Summarize Installed R Libraries}
\description{
Functions to view and summarize installed R libraries and packages on local machine.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/whmacken/libminerWHM}
  \item Report bugs at \url{https://github.com/whmacken/libminerWHM/issues}
}

}
\author{
\strong{Maintainer}: William H. MacKenzie \email{Will.MacKenzie@gov.bc.ca}

}
\keyword{internal}
