% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_summary.R
\name{lib_summary}
\alias{lib_summary}
\title{R Library Summary}
\usage{
lib_summary(sizes = FALSE)
}
\arguments{
\item{sizes}{logical to show library sizes in bytes}
}
\value{
a data.frame containing count of packages in each library
}
\description{
Provides brief summary of installed packages on local machine
}
\examples{
lib_summary()
lib_summary(sizes = TRUE)
}
